Mejora de los estilos, actualizacion del README y arreglo en gitignore

+++ db/schema.rb
@@ -50,7 +50,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_08_01_204100) do
 
   create_table "comments", force: :cascade do |t|
     t.string "content"
-    t.integer "movie_id", null: false
+    t.bigint "movie_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["movie_id"], name: "index_comments_on_movie_id"
@@ -68,8 +68,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_08_01_204100) do
   end
 
   create_table "movie_categories", force: :cascade do |t|
-    t.integer "movie_id", null: false
-    t.integer "category_id", null: false
+    t.bigint "movie_id", null: false
+    t.bigint "category_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["category_id"], name: "index_movie_categories_on_category_id"
